third_party/WebKit/LayoutTests/fast/dom/HTMLLinkElement/link-preload-validity-expected.txt

CONSOLE DEBUG: line 10: Preload triggered for prefetch.wut.com.test/1.gif
CONSOLE WARNING: line 11: <link rel=preload> has an invalid `href` value
CONSOLE WARNING: line 12: <link rel=preload> has an invalid `href` value
CONSOLE WARNING: line 13: <link rel=preload> has an invalid `href` value
CONSOLE DEBUG: line 14: Preload triggered for prefetch.wut.com.test/3.gif
This test verifies that a preload request is sent and logged.
